remove course.js from router folder , rename course.js in query folder
