--- server/src/router/course.js
@@ -1,3 +1,0 @@
-const { getCoursesByCatId } = require('../controllers');
-
-module.exports = { getCoursesByCatId };
